[network] Panic if network identity doesn't match onchain
It's possible that a user could set their network identity onchain and
then continue communicating on the network with the wrong network
identity.  This causes a strange 1 way communication, which makes it
very confusing.  If we panic, we'll prevent communication to the invalid
network key and restart to fix it.

Closes: #7217